fix: Handle paths more correctly
The previous handling had problems with handling paths that ended in
files instead of directories.
This has been resolved as of now.

The method for feeding data to the program now is a struct instead of
strings (and array of strings).
The struct requires:

- The starting location
- The key to be lookedup
- The level to be searched.

For now system is still not functional as I have yet to find what
`$(prefix)` means.
Should be coming soon, (a guess)

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>